backend/requirements.txt
fastapi
uvicorn
pdfplumber
PyPDF2
pyyaml